requirements.txt
fastapi
uvicorn
requests
redis
matplotlib
python-dotenv
black
isort
pytest
